feat(#33): get rooms tests
